change options usage in Appium service
