+++ packages/wdio-appium-service/src/launcher.js
@@ -1,46 +1,47 @@
 import logger from '@wdio/logger'
 import { spawn } from 'child_process'
 import { createWriteStream, ensureFileSync } from 'fs-extra'
-import { paramCase } from 'param-case'
 import { promisify } from 'util'
-import getFilePath from './utils/getFilePath'
+import { getFilePath, getAppiumCommand, cliArgsFromKeyValue } from './utils'
 
 const log = logger('@wdio/appium-service')
 const DEFAULT_LOG_FILENAME = 'appium.txt'
 
 export default class AppiumLauncher {
-    constructor() {
-        this.logPath = null
-        this.command = ''
+    constructor(options, caps, config) {
+        this.options = options
+        this.logPath = options.logPath || config.outputDir
+        this.command = options.command
         this.appiumArgs = []
+
+        /**
+         * Windows expects node to be explicitely set as command and appium
+         * module path as it's first argument
+         */
+        if (!this.command) {
+            this.command = 'node'
+            this.appiumArgs.push(getAppiumCommand())
+        }
     }
 
-    async onPrepare(config) {
-        const appiumConfig = config.appium || {}
+    async onPrepare() {
         const isWindows = process.platform === 'win32'
 
-        this.logPath = appiumConfig.logPath || config.outputDir
+        /**
+         * Append remaining arguments
+         */
+        this.appiumArgs.push(...cliArgsFromKeyValue(this.options.args || {}))
 
-        // Set config command
-        if (appiumConfig.command) {
-            this.command = appiumConfig.command
-        } else {
-            // Windows expects node to be explicitely set as command and appium module path as it's first argument
-            this.command = 'node'
-            this.appiumArgs.push(this._getAppiumCommand())
-        }
-
-        // Append remaining arguments
-        this.appiumArgs.push(...this._cliArgsFromKeyValue(appiumConfig.args || {}))
-
-        // Windows needs to be started through `cmd` and the command needs to be an arg
+        /**
+         * Windows needs to be started through `cmd` and the command needs to be an arg
+         */
         if (isWindows) {
             this.appiumArgs.unshift('/c', this.command)
             this.command = 'cmd'
         }
 
         const asyncStartAppium = promisify(this._startAppium)
-        this.process = await asyncStartAppium(this.command, this.appiumArgs, this.waitStartTime)
+        this.process = await asyncStartAppium(this.command, this.appiumArgs)
 
         if (typeof this.logPath === 'string') {
             this._redirectLogStream(this.logPath)
@@ -54,7 +55,7 @@ export default class AppiumLauncher {
         }
     }
 
-    _startAppium(command, args, waitStartTime, callback) {
+    _startAppium(command, args, callback) {
         log.debug(`Will spawn Appium process: ${command} ${args.join(' ')}`)
         let process = spawn(command, args, { stdio: ['ignore', 'pipe', 'pipe'] })
         let error
@@ -66,7 +67,9 @@ export default class AppiumLauncher {
             }
         })
 
-        // only capture first error to print it in case Appium failed to start.
+        /**
+         * only capture first error to print it in case Appium failed to start.
+         */
         process.stderr.once('data', err => { error = err })
 
         process.once('exit', (exitCode) => {
@@ -89,46 +92,5 @@ export default class AppiumLauncher {
         const logStream = createWriteStream(logFile, { flags: 'w' })
         this.process.stdout.pipe(logStream)
         this.process.stderr.pipe(logStream)
-    }
-
-    _getAppiumCommand(moduleName = 'appium') {
-        try {
-            return require.resolve(moduleName)
-        } catch (err) {
-            log.error('appium is not installed locally.\n' +
-                'If you use globally installed appium please add\n' +
-                'appium: { command: \'appium\' }\n' +
-                'to your wdio.conf.js!')
-            throw err
-        }
-    }
-
-    _cliArgsFromKeyValue(keyValueArgs) {
-        if (Array.isArray(keyValueArgs)) {
-            return keyValueArgs
-        }
-
-        const cliArgs = []
-        for (let key in keyValueArgs) {
-            const value = keyValueArgs[key]
-            // If the value is false or null the argument is discarded
-            if ((typeof value === 'boolean' && !value) || value === null) {
-                continue
-            }
-
-            cliArgs.push(`--${paramCase(key)}`)
-
-            // Only non-boolean and non-null values are added as option values
-            if (typeof value !== 'boolean' && value !== null) {
-                cliArgs.push(this._sanitizeCliOptionValue(value))
-            }
-        }
-        return cliArgs
-    }
-
-    _sanitizeCliOptionValue(value) {
-        const valueString = String(value)
-        // Encapsulate the value string in single quotes if it contains a white space
-        return /\s/.test(valueString) ? `'${valueString}'` : valueString
     }
 }
